+++ inc/class.acf-controls.php
@@ -33,7 +33,7 @@ class AcfControls extends Singleton {
       add_filter("acf/load_field/type=$field_type", [$this, 'load_translatable_field'], 20);
     }
     // filter field wrapper attributes
-    add_filter("acf/field_wrapper_attributes", [$this, 'acf_field_wrapper_attributes'], 10, 2);
+    // add_filter("acf/field_wrapper_attributes", [$this, 'acf_field_wrapper_attributes'], 10, 2);
     // add hooks for generated translatable fields (type of those will be 'group')
     add_filter("acf/format_value/type=group", [$this, 'format_translatable_field_value'], 11, 3);
     add_filter("acf/render_field/type=group", [$this, 'render_translatable_field'], 5);
@@ -91,7 +91,7 @@ class AcfControls extends Singleton {
         'key' => "{$field['key']}_{$lang}",
         'label' => "{$field['label']} ({$lang})",
         'name' => "{$field['name']}_{$lang}",
-        '_name' => "{$field['_name']}_$lang",
+        '_name' => "$lang",
         // Only the default language of a sub-field should be required
         'required' => $lang === $default_language && $field['required'],
         'is_translatable' => 0,
@@ -190,9 +190,10 @@ class AcfControls extends Singleton {
     return is_array($field) && $field['type'] === 'group' && !empty($field['is_translatable']);
   }
 
-  public function acf_field_wrapper_attributes($atts, $field) {
-    // pre_dump([$atts, $field]);
-    return $atts;
+  public function acf_field_wrapper_attributes($wrapper, $field) {
+    if( $field['_name'] !== 'a_translatable_field_en' ) return $wrapper;
+    pre_dump([$field]);
+    return $wrapper;
   }
   
 }